[CPDLP-1850] Address PR comments
