+++ spec/requests/api/v2/npq_applications_spec.rb
@@ -22,7 +22,7 @@ RSpec.describe "NPQ Applications API", :with_default_schedules, type: :request d
       list << create_list(:npq_application, 2, npq_lead_provider: other_npq_lead_provider, school_urn: "123456", npq_course:, cohort:)
 
       list.flatten.each do |npq_application|
-        AcceptNPQApplication.new(npq_application:).call
+        NPQ::AcceptApplication.new(npq_application:).call
       end
     end
